Part 1 of incognito mode
Split this up into multiple commits since it's more readable but here goes:

Adding the preference for incognito mode (currently not togglable in this commit)

This is the logic from upstream, however as per ushe, there's difference in presentation. This one adds an incognito icon to toolbar. How it's displayed in upstream can work here to with (not too much) effort, but boy this was is easier and lazier

Co-Authored-By: arkon <4098258+arkon@users.noreply.github.com>